Remove chapter 12 exercise 10, adjust all subsequent exercise numbers.
