+++ src/test/kotlin/chapter12/exercises/ex10/listing.kt
@@ -1,0 +1,28 @@
+package chapter12.exercises.ex10
+
+import chapter11.Monad
+import chapter12.CompositePartialOf
+
+interface Listing<F, G> {
+
+    //tag::init1[]
+    fun <F, G> compose(
+        mf: Monad<F>,
+        mg: Monad<G>
+    ): Monad<CompositePartialOf<F, G>> = TODO()
+    //end::init1[]
+}
+
+/*
+//tag::init2[]
+fun <A, B> flatMap(
+    mna: CompositeOf<F, G, A>,
+    f: (A) -> CompositeOf<F, G, B>
+): CompositeOf<F, G, B> =
+    mf.flatMap(mna.fix().value) { na: Kind<G, A> ->
+        mg.flatMap(na) { a: A ->
+            f(a)
+        }
+    }
+//end::init2[]
+*/
