+++ src/test/kotlin/chapter12/exercises/ex12/listing.kt
@@ -1,34 +1,18 @@
 package chapter12.exercises.ex12
 
-import arrow.Kind
-import arrow.syntax.function.curried
-import chapter12.Functor
+import chapter10.ForOption
+import chapter12.ForList
+import chapter12.ForTree
+import chapter12.Traversable
 
-interface Applicative<F> : Functor<F> {
+//tag::init1[]
+fun <A> optionTraversable(): Traversable<ForOption> = TODO()
+//end::init1[]
 
-    override fun <A, B> map(
-        fa: Kind<F, A>,
-        f: (A) -> B
-    ): Kind<F, B> =
-        map2(fa, unit(Unit)) { a, _ -> f(a) }
+//tag::init2[]
+fun <A> listTraversable(): Traversable<ForList> = TODO()
+//end::init2[]
 
-    fun <A, B> apply(
-        fab: Kind<F, (A) -> B>,
-        fa: Kind<F, A>
-    ): Kind<F, B>
-
-    fun <A> unit(a: A): Kind<F, A>
-
-    fun <A, B, C> map2(
-        fa: Kind<F, A>,
-        fb: Kind<F, B>,
-        f: (A, B) -> C
-    ): Kind<F, C> =
-        apply(apply(unit(f.curried()), fa), fb)
-
-    //tag::init1[]
-    fun <K, V> sequence(
-        mkv: Map<K, Kind<F, V>>
-    ): Kind<F, Map<K, V>> = TODO()
-    //end::init1[]
-}
+//tag::init3[]
+fun <A> treeTraversable(): Traversable<ForTree> = TODO()
+//tag::init3[]
