+++ src/test/kotlin/chapter12/exercises/ex13/listing.kt
@@ -1,18 +1,29 @@
 package chapter12.exercises.ex13
 
-import chapter10.ForOption
-import chapter12.ForList
-import chapter12.ForTree
-import chapter12.Traversable
+import arrow.Kind
+import arrow.typeclasses.Applicative
+import chapter12.Functor
 
 //tag::init1[]
-fun <A> optionTraversable(): Traversable<ForOption> = TODO()
+interface Traversable<F> : Functor<F> {
+    //tag::ignore[]
+    fun <G, A, B> traverse(
+        fa: Kind<F, A>,
+        AG: Applicative<G>,
+        f: (A) -> Kind<G, B>
+    ): Kind<G, Kind<F, B>> =
+        sequence(map(fa, f), AG)
+
+    fun <G, A> sequence(
+        fga: Kind<F, Kind<G, A>>,
+        AG: Applicative<G>
+    ): Kind<G, Kind<F, A>> =
+        traverse(fga, AG) { it }
+    //end::ignore[]
+
+    override fun <A, B> map(
+        fa: Kind<F, A>,
+        f: (A) -> B
+    ): Kind<F, B> = TODO()
+}
 //end::init1[]
-
-//tag::init2[]
-fun <A> listTraversable(): Traversable<ForList> = TODO()
-//end::init2[]
-
-//tag::init3[]
-fun <A> treeTraversable(): Traversable<ForTree> = TODO()
-//tag::init3[]
