+++ src/test/kotlin/chapter12/exercises/ex19/listing.kt
@@ -1,23 +1,15 @@
 package chapter12.exercises.ex19
 
-import arrow.Kind
+import chapter11.Monad
 import chapter12.Applicative
-import chapter12.Composite
-import chapter12.CompositeOf
 import chapter12.CompositePartialOf
 import chapter12.Traversable
-import chapter12.fix
 
 //tag::init[]
-fun <F, G> compose(
-    TF: Traversable<F>,
-    TG: Traversable<G>
-): Traversable<CompositePartialOf<F, G>> =
-    object : Traversable<CompositePartialOf<F, G>> {
-        override fun <H, A, B> traverse(
-            fa: CompositeOf<F, G, A>,
-            AH: Applicative<H>,
-            f: (A) -> Kind<H, B>
-        ): Kind<H, CompositeOf<F, G, B>> = TODO()
-    }
+fun <G, H, A> composeM(
+    MG: Monad<G>,
+    MH: Monad<H>,
+    AH: Applicative<H>,
+    TH: Traversable<H>
+): Monad<CompositePartialOf<G, H>> = TODO()
 //end::init[]
