--- src/test/kotlin/chapter12/exercises/ex20/listing.kt
@@ -1,19 +1,0 @@
-package chapter12.exercises.ex20
-
-import arrow.Kind
-import chapter11.Monad
-import chapter12.Applicative
-import chapter12.Composite
-import chapter12.CompositeOf
-import chapter12.CompositePartialOf
-import chapter12.Traversable
-import chapter12.fix
-
-//tag::init[]
-fun <G, H, A> composeM(
-    MG: Monad<G>,
-    MH: Monad<H>,
-    AH: Applicative<H>,
-    TH: Traversable<H>
-): Monad<CompositePartialOf<G, H>> = TODO()
-//end::init[]
