+++ src/test/kotlin/chapter12/solutions/ex18/listing.kt
@@ -1,61 +1,28 @@
 package chapter12.solutions.ex18
 
 import arrow.Kind
-import chapter10.Foldable
 import chapter12.Applicative
-import chapter12.Functor
-import chapter12.Product
-import chapter12.ProductOf
-import chapter12.ProductPartialOf
+import chapter12.Composite
+import chapter12.CompositeOf
+import chapter12.CompositePartialOf
+import chapter12.Traversable
 import chapter12.fix
 
-infix fun <F, G> Applicative<F>.product(
-    ag: Applicative<G>
-): Applicative<ProductPartialOf<F, G>> =
-    object : Applicative<ProductPartialOf<F, G>> {
-        override fun <A, B> apply(
-            fgab: ProductOf<F, G, (A) -> B>,
-            fga: ProductOf<F, G, A>
-        ): ProductOf<F, G, B> {
-            val (fab, gab) = fgab.fix().value
-            val (fa, ga) = fga.fix().value
-            return Product(
-                Pair(
-                    this@product.apply(fab, fa),
-                    ag.apply(gab, ga)
-                )
-            )
-        }
-
-        override fun <A> unit(a: A): ProductOf<F, G, A> =
-            Product(Pair(this@product.unit(a), ag.unit(a)))
+//tag::init[]
+fun <F, G> compose(
+    TF: Traversable<F>,
+    TG: Traversable<G>
+): Traversable<CompositePartialOf<F, G>> =
+    object : Traversable<CompositePartialOf<F, G>> {
+        override fun <H, A, B> traverse(
+            fa: CompositeOf<F, G, A>,
+            AH: Applicative<H>,
+            f: (A) -> Kind<H, B>
+        ): Kind<H, CompositeOf<F, G, B>> =
+            AH.map(
+                TF.traverse(fa.fix().value, AH) { ga: Kind<G, A> ->
+                    TG.traverse(ga, AH) { a: A -> f(a) }
+                }
+            ) { Composite(it) }
     }
-
-interface Traversable<F> : Functor<F>, Foldable<F> {
-
-    fun <G, A> sequence(
-        fga: Kind<F, Kind<G, A>>,
-        AG: Applicative<G>
-    ): Kind<G, Kind<F, A>> =
-        traverse(fga, AG) { it }
-
-    fun <G, A, B> traverse(
-        fa: Kind<F, A>,
-        AG: Applicative<G>,
-        f: (A) -> Kind<G, B>
-    ): Kind<G, Kind<F, B>> =
-        sequence(map(fa, f), AG)
-
-    //tag::init[]
-    fun <G, H, A, B> fuse(
-        ta: Kind<F, A>,
-        AG: Applicative<G>,
-        AH: Applicative<H>,
-        f: (A) -> Kind<G, B>,
-        g: (A) -> Kind<H, B>
-    ): Pair<Kind<G, Kind<F, B>>, Kind<H, Kind<F, B>>> =
-        traverse(ta, AG product AH) { a ->
-            Product(Pair(f(a), g(a)))
-        }.fix().value
-    //end::init[]
-}
+//end::init[]
